requirements.txt
PyMySQL==0.9.3
SQLAlchemy==1.3.10
Flask==1.1.1
PyJWT==1.7.1
